% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{clean_all}
\alias{clean_all}
\title{clean_all}
\usage{
clean_all(files)
}
\arguments{
\item{files}{optional direction for photos, if not given, all the photos in "static/img" are used}
}
\description{
Scale and orient all blog photos
}
\examples{
clean_all()
}
