% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_travel_blog.R
\name{new_travel_blog}
\alias{new_travel_blog}
\title{new_travel_blog}
\usage{
new_travel_blog(path, ...)
}
\arguments{
\item{path}{target directory where to put the new blog}

\item{...}{template input rstudio}
}
\description{
Create a new travel blog
}
