% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{new_trip}
\alias{new_trip}
\title{new_trip}
\usage{
new_trip(trip_name, title, image)
}
\arguments{
\item{trip_name}{short name of the trip (for folders)}

\item{title}{Title of your trip (for display)}

\item{image}{header photo location (you'll can find it at content/trip_name/header.jpg)}
}
\description{
Create a new trip in the blog
}
\examples{
new_trip("NYC_2018", "New York 2018")
}
