% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trip_post.R
\name{new_trip_post}
\alias{new_trip_post}
\title{new_trip_post}
\usage{
new_trip_post(trip_name, title, date = Sys.Date(), author = "me", header,
  tags, place)
}
\arguments{
\item{trip_name}{short name of the trip}

\item{title}{title of the post}

\item{date}{date in standard format "2018-04-01" (default = date of the day)}

\item{author}{author's name (default = "me")}

\item{header}{image for the post}

\item{tags}{tags}
}
\description{
Create a new post for a trip in the blog
}
\examples{
\dontrun{ new_trip_post("NYC_2018", "Brooklyn Heights")}
}
