% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_gallery.R
\name{opendir}
\alias{opendir}
\title{open_dir}
\usage{
opendir(dir = getwd())
}
\arguments{
\item{dir}{optional directory to open, current directory by default}
}
\description{
To open the post directory to put the photos
}
