% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{orient_all}
\alias{orient_all}
\title{orient_all}
\usage{
orient_all(files)
}
\arguments{
\item{files}{optional direction for photos, if not given, all the photos in "static/img" are used}
}
\description{
Orient all blog photos
}
\examples{
orient_all()
}
