% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{redim_all}
\alias{redim_all}
\title{redim_all}
\usage{
redim_all(files)
}
\arguments{
\item{files}{optional direction for photos, if not given, all the photos in "static/img" are used}
}
\description{
Scale all blog photos
}
\examples{
redim_all()
}
